Changed to CLang, fixed text rendering - starting new BSP

+++ src/rendering/shader.c
@@ -101,7 +101,7 @@ bool Shader_Load(shader_t* shader, const char* assetPath, const char* vertPath, 
   shader->uSpecularLoc = glGetUniformLocation(shader->program, "uSpecular");
   shader->uShininessLoc = glGetUniformLocation(shader->program, "uShininess");
 
-  for (size_t i = 0; i < MAX_FORWARD_LIGHTS; i++){
+  for (int i = 0; i < MAX_FORWARD_LIGHTS; i++){
     char uniform[128];
 
     // Position
